+++ lib/src/api/Logger/logger.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/foundation.dart';
+
+class Logger {
+  static void log(String message) {
+    String timestamp = DateTime.now().toIso8601String();
+    if (kDebugMode) {
+      print("$timestamp | giphy_api | $message");
+    }
+  }
+}