+++ lib/src/ui/Platforms/platforms_web.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/foundation.dart';
+
+class Platforms {
+  static const bool isAndroid = false;
+  static const bool isIOS = false;
+  static const bool isWeb = kIsWeb;
+  static const bool isDesktop = false;
+  static const bool isMacOS = false;
+  static const bool isWindows = false;
+  static const bool isLinux = false;
+  static const bool isFuchsia = false;
+}