+++ lib/src/ui/View/Emojis/emojis.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_giphy_picker/giphy_api.dart';
+import 'package:flutter_giphy_picker/src/ui/Functions/parse_exception.dart';
 import 'package:flutter_giphy_picker/src/ui/View/Emojis/emojis_variations.dart';
 import 'package:flutter_giphy_picker/src/ui/View/Emojis/select_emoji.dart';
 import 'package:flutter_giphy_picker/src/ui/giphy_config.dart';
@@ -90,7 +91,7 @@ class _EmojisView extends State<EmojisView> {
             )
           : error
               ? Center(
-                  child: Text(exception.toString()),
+                  child: Text(parseException(exception)),
                 )
               : MasonryGridView.extent(
                   maxCrossAxisExtent: 250,